@@ -29,11 +29,11 @@ import org.wildfly.transaction.client.spi.SimpleTransactionControl;
 import org.xnio.IoUtils;
 
 import javax.net.ssl.SSLContext;
-import javax.transaction.HeuristicMixedException;
-import javax.transaction.HeuristicRollbackException;
-import javax.transaction.RollbackException;
-import javax.transaction.Status;
-import javax.transaction.SystemException;
+import jakarta.transaction.HeuristicMixedException;
+import jakarta.transaction.HeuristicRollbackException;
+import jakarta.transaction.RollbackException;
+import jakarta.transaction.Status;
+import jakarta.transaction.SystemException;
 import javax.transaction.xa.Xid;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
